refactor(shopper): CHECKOUT-3791 Move customer strategy options interface
