+++ src/customer/strategies/amazon/amazon-pay-customer-strategy.ts
@@ -1,12 +1,13 @@
-import { CheckoutStore, InternalCheckoutSelectors} from '../../checkout';
-import { InvalidArgumentError, MissingDataError, MissingDataErrorType, NotImplementedError, NotInitializedError, NotInitializedErrorType, StandardError } from '../../common/error/errors';
-import { PaymentMethod, PaymentMethodActionCreator } from '../../payment';
-import { AmazonPayLoginButton, AmazonPayScriptLoader, AmazonPayWidgetError, AmazonPayWindow } from '../../payment/strategies/amazon-pay';
-import { RemoteCheckoutActionCreator, RemoteCheckoutRequestSender } from '../../remote-checkout';
-import CustomerCredentials from '../customer-credentials';
-import { CustomerInitializeOptions, CustomerRequestOptions } from '../customer-request-options';
+import { CheckoutStore, InternalCheckoutSelectors} from '../../../checkout';
+import { InvalidArgumentError, MissingDataError, MissingDataErrorType, NotImplementedError, NotInitializedError, NotInitializedErrorType } from '../../../common/error/errors';
+import { PaymentMethod, PaymentMethodActionCreator } from '../../../payment';
+import { AmazonPayLoginButton, AmazonPayScriptLoader, AmazonPayWindow } from '../../../payment/strategies/amazon-pay';
+import { RemoteCheckoutActionCreator, RemoteCheckoutRequestSender } from '../../../remote-checkout';
+import CustomerCredentials from '../../customer-credentials';
+import { CustomerInitializeOptions, CustomerRequestOptions } from '../../customer-request-options';
+import CustomerStrategy from '../customer-strategy';
 
-import CustomerStrategy from './customer-strategy';
+import AmazonPayCustomerInitializeOptions from './amazon-pay-customer-initialize-options';
 
 export default class AmazonPayCustomerStrategy implements CustomerStrategy {
     private _paymentMethod?: PaymentMethod;
@@ -120,39 +121,6 @@ export default class AmazonPayCustomerStrategy implements CustomerStrategy {
                 this._remoteCheckoutRequestSender.trackAuthorizationEvent();
             });
     }
-}
-
-/**
- * A set of options that are required to initialize the customer step of
- * checkout to support Amazon Pay.
- *
- * When AmazonPay is initialized, a sign-in button will be inserted into the
- * DOM. When the customer clicks on it, they will be redirected to Amazon to
- * sign in.
- */
-export interface AmazonPayCustomerInitializeOptions {
-    /**
-     * The ID of a container which the sign-in button should insert into.
-     */
-    container: string;
-
-    /**
-     * The colour of the sign-in button.
-     */
-    color?: 'Gold' | 'LightGray' | 'DarkGray';
-
-    /**
-     * The size of the sign-in button.
-     */
-    size?: 'small' | 'medium' | 'large' | 'x-large';
-
-    /**
-     * A callback that gets called if unable to initialize the widget or select
-     * one of the address options provided by the widget.
-     *
-     * @param error - The error object describing the failure.
-     */
-    onError?(error: AmazonPayWidgetError | StandardError): void;
 }
 
 interface AuthorizationOptions {
