+++ src/customer/strategies/amazon/index.ts
@@ -1,0 +1,2 @@
+export { default as AmazonPayCustomerStrategy } from './amazon-pay-customer-strategy';
+export { default as AmazonPayCustomerInitializeOptions } from './amazon-pay-customer-initialize-options';
