+++ src/customer/strategies/braintree/braintree-visacheckout-customer-initialize-options.ts
@@ -1,0 +1,4 @@
+export default interface BraintreeVisaCheckoutCustomerInitializeOptions {
+    container: string;
+    onError?(error: Error): void;
+}
