+++ src/customer/strategies/braintree/index.ts
@@ -1,0 +1,2 @@
+export { default as BraintreeVisaCheckoutCustomerStrategy } from './braintree-visacheckout-customer-strategy';
+export { default as BraintreeVisaCheckoutCustomerInitializeOptions } from './braintree-visacheckout-customer-initialize-options';
