+++ src/customer/strategies/chasepay/chasepay-customer-initialize-options.ts
@@ -1,0 +1,3 @@
+export default interface ChasePayCustomerInitializeOptions {
+    container: string;
+}
