+++ src/customer/strategies/default/index.ts
@@ -1,0 +1,1 @@
+export { default as DefaultCustomerStrategy } from './default-customer-strategy';
