+++ src/customer/strategies/index.ts
@@ -1,9 +1,1 @@
-export { default as AmazonPayCustomerStrategy, AmazonPayCustomerInitializeOptions } from './amazon-pay-customer-strategy';
 export { default as CustomerStrategy } from './customer-strategy';
-export { default as DefaultCustomerStrategy } from './default-customer-strategy';
-export { default as BraintreeVisaCheckoutCustomerStrategy, BraintreeVisaCheckoutCustomerInitializeOptions } from './braintree-visacheckout-customer-strategy';
-export { default as ChasePayCustomerStrategy, ChasePayCustomerInitializeOptions } from './chasepay-customer-strategy';
-export { default as SquareCustomerStrategy } from './square-customer-strategy';
-export { default as MasterpassCustomerInitializeOptions} from './masterpass-customer-initialize-options';
-export { default as MasterpassCustomerStrategy } from './masterpass-customer-strategy';
-export { GooglePayCustomerStrategy, GooglePayCustomerInitializeOptions } from './googlepay';
