+++ src/customer/strategies/masterpass/index.ts
@@ -1,0 +1,2 @@
+export { default as MasterpassCustomerInitializeOptions } from './masterpass-customer-initialize-options';
+export { default as MasterpassCustomerStrategy } from './masterpass-customer-strategy';
