+++ src/customer/strategies/square/index.ts
@@ -1,0 +1,1 @@
+export { default as SquareCustomerStrategy } from './square-customer-strategy';
