+++ src/customer/strategies/square/square-customer-strategy.ts
@@ -1,10 +1,10 @@
-import { CheckoutStore, InternalCheckoutSelectors } from '../../checkout';
-import { NotImplementedError} from '../../common/error/errors';
-import { RemoteCheckoutActionCreator } from '../../remote-checkout';
-import CustomerCredentials from '../customer-credentials';
-import { CustomerInitializeOptions, CustomerRequestOptions } from '../customer-request-options';
+import { CheckoutStore, InternalCheckoutSelectors } from '../../../checkout';
+import { NotImplementedError} from '../../../common/error/errors';
+import { RemoteCheckoutActionCreator } from '../../../remote-checkout';
+import CustomerCredentials from '../../customer-credentials';
+import { CustomerInitializeOptions, CustomerRequestOptions } from '../../customer-request-options';
 
-import CustomerStrategy from './customer-strategy';
+import CustomerStrategy from '../customer-strategy';
 
 export default class SquareCustomerStrategy implements CustomerStrategy {
